requirements.txt
langchain
pydantic
openai
pandas
langchain-community
logparser3
tiktoken
faiss-cpu
streamlit